Basic preferences page established
