+++ src/DigitalProduction.Maui.sln
@@ -1,15 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
-VisualStudioVersion = 18.5.11716.220 stable
+VisualStudioVersion = 18.5.11716.220
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalProduction.Demo", "Demo\DigitalProduction.Demo.csproj", "{61056E84-6CBE-471C-8A4D-C12F226CDBF0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalProduction.Maui", "Maui Library\DigitalProduction.Maui.csproj", "{129D30B3-904B-4259-8CD4-5C9A70621B33}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalProduction.Maui.Images", "Image Library\DigitalProduction.Maui.Images.csproj", "{25FB275F-B23C-4355-A866-88D1C3C9884D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Unit Tests", "Unit Tests\Unit Tests.csproj", "{F6F9AE27-9710-4D95-A4E7-C6E3950F3BC6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,10 +28,6 @@ Global
 		{25FB275F-B23C-4355-A866-88D1C3C9884D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25FB275F-B23C-4355-A866-88D1C3C9884D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25FB275F-B23C-4355-A866-88D1C3C9884D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F6F9AE27-9710-4D95-A4E7-C6E3950F3BC6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F6F9AE27-9710-4D95-A4E7-C6E3950F3BC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F6F9AE27-9710-4D95-A4E7-C6E3950F3BC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F6F9AE27-9710-4D95-A4E7-C6E3950F3BC6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
